refactoring for rubocop and codeclimate
